requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
crispy-bootstrap5==0.6
cryptography==37.0.2
dj-database-url==0.5.0
Django==4.0.6
django-crispy-forms==1.14.0
django-heroku==0.3.1
dominate==2.6.0
Flask==2.1.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.0
MarkupSafe==2.1.1
numpy==1.22.4
outcome==1.2.0
pandas==1.4.2
Pillow==9.2.0
portalocker==2.5.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyperclip==1.8.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pywin32==304
requests==2.28.1
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
sqlparse==0.4.2
trio==0.21.0
trio-websocket==0.9.2
twilio==7.9.3
tzdata==2022.1
urllib3==1.26.10
visitor==0.1.3
waitress==2.1.2
Werkzeug==2.1.2
whitenoise==6.2.0
wsproto==1.1.0
WTForms==3.0.1